+++ CHANGELOG.md
@@ -11,6 +11,10 @@
 
 ##### Enhancements
 
+* FunctionBodyLengthRule now does not count comment lines.  
+  [Marcelo Fabri](https://github.com/marcelofabri)
+  [#258](https://github.com/realm/SwiftLint/issues/258)
+
 * `ConfigurableRule` protocol allows for improved rule configuration. See
   `CONTRIBUTING` for more details.  
   [Scott Hoyt](https://github.com/scottrhoyt)